msm: camera: Changes to accommodate 32-bit Arch
This driver is using uint64_t to store pointers. On 64 bit uint64_t
and pointer is of the same size, but on the 32 bit system pointers are
smaller. Casting uint64_t to pointer generates warning on 32 bit
platforms.
Fix this by storing pointers as uintptr_t.

On a 32 bit system, pointer size is 32 bits, but userspace is passing
uint64_t value. Casting uint64_t to pointer generates a compilation
warning.
Fix this by first casting uint64_t to uintptr_t then to void ptr.

+++ drivers/media/platform/msm/camera/cam_req_mgr/cam_mem_mgr.c
@@ -25,11 +25,11 @@
 static struct cam_mem_table tbl;
 
 static int cam_mem_util_map_cpu_va(struct ion_handle *hdl,
-	uint64_t *vaddr,
+	uintptr_t *vaddr,
 	size_t *len)
 {
 	*vaddr = (uintptr_t)ion_map_kernel(tbl.client, hdl);
-	if (IS_ERR_OR_NULL((void *)*vaddr)) {
+	if (IS_ERR_OR_NULL((void *)(uintptr_t)(*vaddr))) {
 		CAM_ERR(CAM_MEM, "kernel map fail");
 		return -ENOSPC;
 	}
@@ -183,12 +183,12 @@ handle_mismatch:
 }
 EXPORT_SYMBOL(cam_mem_get_io_buf);
 
-int cam_mem_get_cpu_buf(int32_t buf_handle, uint64_t *vaddr_ptr, size_t *len)
+int cam_mem_get_cpu_buf(int32_t buf_handle, uintptr_t *vaddr_ptr, size_t *len)
 {
 	int rc = 0;
 	int idx;
 	struct ion_handle *ion_hdl = NULL;
-	uint64_t kvaddr = 0;
+	uintptr_t kvaddr = 0;
 	size_t klen = 0;
 
 	if (!buf_handle || !vaddr_ptr || !len)
@@ -288,7 +288,7 @@ int cam_mem_mgr_cache_ops(struct cam_mem_cache_ops_cmd *cmd)
 
 		rc = msm_ion_do_cache_op(tbl.client,
 				tbl.bufq[idx].i_hdl,
-				(void *)tbl.bufq[idx].vaddr,
+				(void *)(uintptr_t)tbl.bufq[idx].vaddr,
 				tbl.bufq[idx].len,
 				ion_cache_ops);
 		if (rc)
@@ -926,7 +926,7 @@ int cam_mem_mgr_request_mem(struct cam_mem_mgr_request_desc *inp,
 	int rc = 0;
 	uint32_t heap_id;
 	int32_t ion_flag = 0;
-	uint64_t kvaddr;
+	uintptr_t kvaddr;
 	dma_addr_t iova = 0;
 	size_t request_len = 0;
 	uint32_t mem_handle;
